pyproject.toml
[project]
name = "assignment-evaluator"
version = "0.1.0"
description = "Flask API for evaluating assignment repos"
requires-python = ">=3.10"
dependencies = [
    "flask",
    "python-dotenv",
    "langchain-groq",
    "langchain-core",
    "langgraph",
    "GitPython",
]

[project.scripts]
app = "run_workflow:app"
